split packetrefs into separate modules
